+++ kumihan_formatter/core/security/runtime_monitor.py
@@ -1,0 +1,127 @@
+"""
+脆弱性スキャナー - ランタイム監視システム
+
+実行時セキュリティ監視機能
+vulnerability_scanner.pyから分離（Issue: 巨大ファイル分割 - 960行→200行程度）
+"""
+
+import threading
+import time
+from datetime import datetime, timezone
+from typing import Any, Dict, List, Optional
+
+from .vuln_types import RiskLevel, ScanResult, ScannerConfig, VulnerabilityType
+
+
+class RuntimeMonitor:
+    """実行時セキュリティ監視"""
+
+    def __init__(self, config: ScannerConfig, logger: Any):
+        self.config = config
+        self.logger = logger
+        self._monitoring = False
+        self._monitor_thread: Optional[threading.Thread] = None
+        self._suspicious_activities: List[Dict[str, Any]] = []
+
+    def start_monitoring(self) -> None:
+        """監視開始"""
+        if self._monitoring:
+            return
+
+        self._monitoring = True
+        self._monitor_thread = threading.Thread(target=self._monitor_loop, daemon=True)
+        self._monitor_thread.start()
+
+    def stop_monitoring(self) -> None:
+        """監視停止"""
+        self._monitoring = False
+        if self._monitor_thread and self._monitor_thread.is_alive():
+            self._monitor_thread.join(timeout=5)
+
+    def get_security_events(self) -> List[ScanResult]:
+        """セキュリティイベント取得"""
+        results = []
+
+        for activity in self._suspicious_activities:
+            results.append(
+                ScanResult(
+                    vulnerability_type=VulnerabilityType.RUNTIME_BEHAVIOR,
+                    risk_level=activity.get("risk_level", RiskLevel.MEDIUM),
+                    title=activity.get("title", "Suspicious runtime activity"),
+                    description=activity.get("description", ""),
+                    location=activity.get("location", "runtime"),
+                    recommendation=activity.get(
+                        "recommendation", "Investigate the activity"
+                    ),
+                    details=activity,
+                )
+            )
+
+        return results
+
+    def _monitor_loop(self) -> None:
+        """監視メインループ"""
+        while self._monitoring:
+            try:
+                # CPU・メモリ使用量監視
+                self._check_resource_usage()
+
+                # ネットワーク活動監視
+                self._check_network_activity()
+
+                # ファイルシステムアクセス監視
+                self._check_file_access()
+
+                time.sleep(5)  # 5秒間隔
+
+            except Exception as e:
+                self.logger.error(f"Runtime monitoring error: {e}")
+
+    def _check_resource_usage(self) -> None:
+        """リソース使用量チェック"""
+        try:
+            import psutil
+
+            # CPU使用率が90%超過
+            cpu_percent = psutil.cpu_percent(interval=1)
+            if cpu_percent > 90:
+                self._suspicious_activities.append(
+                    {
+                        "title": "High CPU usage detected",
+                        "description": f"CPU usage: {cpu_percent}%",
+                        "risk_level": RiskLevel.MEDIUM,
+                        "timestamp": datetime.now(timezone.utc).isoformat(),
+                        "cpu_percent": cpu_percent,
+                    }
+                )
+
+            # メモリ使用率が85%超過
+            memory = psutil.virtual_memory()
+            if memory.percent > 85:
+                self._suspicious_activities.append(
+                    {
+                        "title": "High memory usage detected",
+                        "description": f"Memory usage: {memory.percent}%",
+                        "risk_level": RiskLevel.MEDIUM,
+                        "timestamp": datetime.now(timezone.utc).isoformat(),
+                        "memory_percent": memory.percent,
+                    }
+                )
+
+        except ImportError:
+            pass
+        except Exception as e:
+            self.logger.error(f"Resource usage check error: {e}")
+
+    def _check_network_activity(self) -> None:
+        """ネットワーク活動チェック"""
+        # 実装簡略化のため基本的なチェックのみ
+        pass
+
+    def _check_file_access(self) -> None:
+        """ファイルアクセスチェック"""
+        # 実装簡略化のため基本的なチェックのみ
+        pass
+
+
+__all__ = ["RuntimeMonitor"]
