+++ kumihan_formatter/core/security/vulnerability_scanner.py
@@ -1,5 +1,18 @@
 """
-脆弱性スキャナー: セキュリティシステムの中核コンポーネント
+脆弱性スキャナー: セキュリティシステムの中核コンポーネント - 統合インターフェース
+
+巨大ファイル分割完了（Issue: 960行→6ファイル分離）
+=======================================================
+
+分割結果:
+- vuln_types.py: 基本型・データクラス（110行）
+- dependency_scanner.py: 依存関係脆弱性検査（206行）  
+- code_pattern_scanner.py: コードパターン脆弱性検出（165行）
+- runtime_monitor.py: 実行時異常行動監視（110行）
+- cve_database.py: CVEデータベース統合（73行）
+- vulnerability_scanner.py: 統合インターフェース（本ファイル, 200行以下）
+
+合計削減効果: 960行 → 864行（96行削減 + 責任分離達成）
 
 複合的なセキュリティ脆弱性検出システム:
 - 依存関係脆弱性検査
@@ -11,646 +24,29 @@
 Claude直接実装 - セキュリティクリティカルなため
 """
 
-import ast
 import hashlib
 import json
 import os
-import re
-import threading
 import time
-from dataclasses import dataclass, field
 from datetime import datetime, timezone
-from enum import Enum
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Union
 
 from kumihan_formatter.core.utilities.logger import get_logger
 
-
-class RiskLevel(Enum):
-    """リスク評価レベル"""
-
-    CRITICAL = "CRITICAL"
-    HIGH = "HIGH"
-    MEDIUM = "MEDIUM"
-    LOW = "LOW"
-    INFO = "INFO"
-
-
-class VulnerabilityType(Enum):
-    """脆弱性タイプ分類"""
-
-    DEPENDENCY = "dependency"
-    CODE_PATTERN = "code_pattern"
-    RUNTIME_BEHAVIOR = "runtime_behavior"
-    CVE_MATCH = "cve_match"
-    CONFIGURATION = "configuration"
-
-
-@dataclass
-class CVERecord:
-    """CVE脆弱性レコード"""
-
-    cve_id: str
-    description: str
-    severity: str
-    score: float
-    affected_versions: List[str]
-    fixed_in: Optional[str] = None
-    published_date: Optional[str] = None
-    references: List[str] = field(default_factory=list)
-
-
-@dataclass
-class ScanResult:
-    """スキャン結果データクラス"""
-
-    vulnerability_type: VulnerabilityType
-    risk_level: RiskLevel
-    title: str
-    description: str
-    location: str
-    recommendation: str
-    details: Dict[str, Any] = field(default_factory=dict)
-    cve_records: List[CVERecord] = field(default_factory=list)
-    timestamp: str = field(
-        default_factory=lambda: datetime.now(timezone.utc).isoformat()
-    )
-
-
-@dataclass
-class SecurityReport:
-    """セキュリティレポート"""
-
-    scan_id: str
-    timestamp: str
-    project_path: str
-    scan_results: List[ScanResult]
-    summary: Dict[str, Any]
-    performance_metrics: Dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass
-class ScannerConfig:
-    """スキャナー設定"""
-
-    enable_dependency_scan: bool = True
-    enable_code_pattern_scan: bool = True
-    enable_runtime_monitor: bool = True
-    enable_cve_checks: bool = True
-    max_file_size: int = 10 * 1024 * 1024  # 10MB
-    scan_timeout: int = 300  # 5分
-    parallel_workers: int = 4
-    exclude_paths: List[str] = field(
-        default_factory=lambda: [
-            ".git",
-            "__pycache__",
-            ".pytest_cache",
-            "node_modules",
-            "venv",
-            ".venv",
-        ]
-    )
-    exclude_extensions: List[str] = field(
-        default_factory=lambda: [".pyc", ".pyo", ".so", ".dll", ".exe", ".bin"]
-    )
-
-
-class DependencyScanner:
-    """依存関係脆弱性スキャナー"""
-
-    def __init__(self, config: ScannerConfig, logger: Any):
-        self.config = config
-        self.logger = logger
-
-    def scan_requirements_file(self, file_path: Path) -> List[ScanResult]:
-        """requirements.txtファイルのスキャン"""
-        results: List[ScanResult] = []
-
-        try:
-            if not file_path.exists():
-                return results
-
-            content = file_path.read_text(encoding="utf-8")
-            lines = content.strip().split("\n")
-
-            for line_no, line in enumerate(lines, 1):
-                line = line.strip()
-                if not line or line.startswith("#"):
-                    continue
-
-                # パッケージ名とバージョン抽出
-                package_info = self._parse_requirement_line(line)
-                if package_info:
-                    vulnerability = self._check_dependency_vulnerability(package_info)
-                    if vulnerability:
-                        vulnerability.location = f"{file_path}:{line_no}"
-                        results.append(vulnerability)
-
-        except Exception as e:
-            self.logger.error(f"Requirements file scan error: {e}")
-
-        return results
-
-    def scan_pyproject_toml(self, file_path: Path) -> List[ScanResult]:
-        """pyproject.tomlファイルのスキャン"""
-        results: List[ScanResult] = []
-
-        try:
-            import tomllib
-
-            if not file_path.exists():
-                return results
-
-            with open(file_path, "rb") as f:
-                data = tomllib.load(f)
-
-            dependencies = []
-
-            # [project.dependencies]
-            if "project" in data and "dependencies" in data["project"]:
-                dependencies.extend(data["project"]["dependencies"])
-
-            # [build-system.requires]
-            if "build-system" in data and "requires" in data["build-system"]:
-                dependencies.extend(data["build-system"]["requires"])
-
-            for dep in dependencies:
-                package_info = self._parse_requirement_line(dep)
-                if package_info:
-                    vulnerability = self._check_dependency_vulnerability(package_info)
-                    if vulnerability:
-                        vulnerability.location = str(file_path)
-                        results.append(vulnerability)
-
-        except ImportError:
-            self.logger.warning("tomllib not available, skipping pyproject.toml scan")
-        except Exception as e:
-            self.logger.error(f"pyproject.toml scan error: {e}")
-
-        return results
-
-    def _parse_requirement_line(self, line: str) -> Optional[Dict[str, str]]:
-        """要求行のパース"""
-        try:
-            # 基本的な形式: package==1.2.3, package>=1.0.0
-            match = re.match(r"^([a-zA-Z0-9_-]+)\s*([><=!]+)\s*([0-9.]+)", line)
-            if match:
-                return {
-                    "name": match.group(1).lower(),
-                    "operator": match.group(2),
-                    "version": match.group(3),
-                }
-
-            # パッケージ名のみ
-            match = re.match(r"^([a-zA-Z0-9_-]+)$", line)
-            if match:
-                return {"name": match.group(1).lower(), "operator": "", "version": ""}
-
-        except Exception:
-            pass
-
-        return None
-
-    def _check_dependency_vulnerability(
-        self, package_info: Dict[str, str]
-    ) -> Optional[ScanResult]:
-        """依存関係脆弱性チェック"""
-        package_name = package_info["name"]
-        version = package_info.get("version", "")
-
-        # 既知の脆弱性パッケージデータベース（簡略版）
-        vulnerable_packages: Dict[str, Dict[str, Any]] = {
-            "django": {
-                "vulnerable_versions": ["<3.2.16", "<4.0.8", "<4.1.3"],
-                "description": (
-                    "Django has known security vulnerabilities in older versions"
-                ),
-                "risk_level": RiskLevel.HIGH,
-            },
-            "flask": {
-                "vulnerable_versions": ["<2.2.0"],
-                "description": (
-                    "Flask has known security vulnerabilities in older versions"
-                ),
-                "risk_level": RiskLevel.MEDIUM,
-            },
-            "requests": {
-                "vulnerable_versions": ["<2.31.0"],
-                "description": (
-                    "Requests library has SSL certificate verification issues"
-                ),
-                "risk_level": RiskLevel.MEDIUM,
-            },
-            "pillow": {
-                "vulnerable_versions": ["<9.3.0"],
-                "description": "Pillow has image processing vulnerabilities",
-                "risk_level": RiskLevel.HIGH,
-            },
-            "pyyaml": {
-                "vulnerable_versions": ["<6.0"],
-                "description": "PyYAML has unsafe loading vulnerabilities",
-                "risk_level": RiskLevel.CRITICAL,
-            },
-        }
-
-        if package_name in vulnerable_packages:
-            vuln_data = vulnerable_packages[package_name]
-
-            # バージョンチェック（簡略化）
-            is_vulnerable = False
-            if version:
-                # 簡単なバージョン比較
-                for vuln_version in vuln_data["vulnerable_versions"]:
-                    if self._version_matches(version, vuln_version):
-                        is_vulnerable = True
-                        break
-            else:
-                # バージョン不明の場合は脆弱性ありと仮定
-                is_vulnerable = True
-
-            if is_vulnerable:
-                return ScanResult(
-                    vulnerability_type=VulnerabilityType.DEPENDENCY,
-                    risk_level=vuln_data["risk_level"],
-                    title=f"Vulnerable dependency: {package_name}",
-                    description=vuln_data["description"],
-                    location="",  # 後で設定
-                    recommendation=f"Update {package_name} to a secure version",
-                    details={
-                        "package": package_name,
-                        "current_version": version,
-                        "vulnerable_versions": vuln_data["vulnerable_versions"],
-                    },
-                )
-
-        return None
-
-    def _version_matches(self, version: str, pattern: str) -> bool:
-        """バージョンパターンマッチング（簡略化）"""
-        try:
-            if pattern.startswith("<"):
-                pattern_version = pattern[1:]
-                return version < pattern_version
-            elif pattern.startswith("<="):
-                pattern_version = pattern[2:]
-                return version <= pattern_version
-            elif pattern.startswith(">"):
-                pattern_version = pattern[1:]
-                return version > pattern_version
-            elif pattern.startswith(">="):
-                pattern_version = pattern[2:]
-                return version >= pattern_version
-            else:
-                return version == pattern
-        except Exception:
-            return True  # 不明な場合は脆弱性ありと仮定
-
-
-class CodePatternScanner:
-    """コードパターン脆弱性スキャナー"""
-
-    # 危険な関数・パターン定義
-    DANGEROUS_FUNCTIONS: Dict[str, Dict[str, Any]] = {
-        "eval": {
-            "risk_level": RiskLevel.CRITICAL,
-            "description": "eval() function allows arbitrary code execution",
-            "recommendation": "Use ast.literal_eval() or avoid dynamic evaluation",
-        },
-        "exec": {
-            "risk_level": RiskLevel.CRITICAL,
-            "description": "exec() function allows arbitrary code execution",
-            "recommendation": "Avoid dynamic code execution",
-        },
-        "compile": {
-            "risk_level": RiskLevel.HIGH,
-            "description": "compile() can execute arbitrary code",
-            "recommendation": "Validate input before compilation",
-        },
-        "__import__": {
-            "risk_level": RiskLevel.HIGH,
-            "description": "Dynamic imports can be security risks",
-            "recommendation": "Use static imports when possible",
-        },
-    }
-
-    # セキュリティパターン
-    SECURITY_PATTERNS: List[Dict[str, Any]] = [
-        {
-            "pattern": r"subprocess\.call.*shell\s*=\s*True",
-            "risk_level": RiskLevel.CRITICAL,
-            "title": "Command injection vulnerability",
-            "description": "subprocess with shell=True can lead to command injection",
-            "recommendation": "Use subprocess without shell=True and validate inputs",
-        },
-        {
-            "pattern": r"os\.system\s*\(",
-            "risk_level": RiskLevel.CRITICAL,
-            "title": "OS command execution",
-            "description": "os.system() can execute arbitrary system commands",
-            "recommendation": ("Use subprocess module with proper input validation"),
-        },
-        {
-            "pattern": r"pickle\.loads?\s*\(",
-            "risk_level": RiskLevel.HIGH,
-            "title": "Unsafe deserialization",
-            "description": (
-                "pickle.load() can execute arbitrary code during deserialization"
-            ),
-            "recommendation": "Use safer serialization formats like JSON",
-        },
-        {
-            "pattern": r"yaml\.load\s*\([^,)]*\)",
-            "risk_level": RiskLevel.HIGH,
-            "title": "Unsafe YAML loading",
-            "description": "yaml.load() without Loader parameter is unsafe",
-            "recommendation": "Use yaml.safe_load() or yaml.load() with SafeLoader",
-        },
-        {
-            "pattern": r"(password|secret|key|token)\s*=\s*['\"][^'\"]{8,}['\"]",
-            "risk_level": RiskLevel.MEDIUM,
-            "title": "Hardcoded credentials",
-            "description": "Hardcoded credentials found in source code",
-            "recommendation": "Use environment variables or secure configuration",
-        },
-        {
-            "pattern": r"\.format\([^)]*\{.*\}[^)]*\)",
-            "risk_level": RiskLevel.LOW,
-            "title": "Potential format string vulnerability",
-            "description": "Dynamic string formatting can be vulnerable",
-            "recommendation": "Use parameterized queries or validate format strings",
-        },
-    ]
-
-    def __init__(self, config: ScannerConfig, logger: Any):
-        self.config = config
-        self.logger = logger
-
-    def scan_file(self, file_path: Path) -> List[ScanResult]:
-        """単一ファイルのスキャン"""
-        results: List[ScanResult] = []
-
-        try:
-            if file_path.stat().st_size > self.config.max_file_size:
-                return results
-
-            content = file_path.read_text(encoding="utf-8")
-
-            # AST解析による関数呼び出しチェック
-            ast_results = self._scan_ast(file_path, content)
-            results.extend(ast_results)
-
-            # 正規表現パターンチェック
-            pattern_results = self._scan_patterns(file_path, content)
-            results.extend(pattern_results)
-
-        except Exception as e:
-            self.logger.error(f"File scan error for {file_path}: {e}")
-
-        return results
-
-    def _scan_ast(self, file_path: Path, content: str) -> List[ScanResult]:
-        """AST解析によるスキャン"""
-        results: List[ScanResult] = []
-
-        try:
-            tree = ast.parse(content)
-
-            for node in ast.walk(tree):
-                if isinstance(node, ast.Call):
-                    func_name = self._get_function_name(node.func)
-                    if func_name in self.DANGEROUS_FUNCTIONS:
-                        func_data = self.DANGEROUS_FUNCTIONS[func_name]
-
-                        results.append(
-                            ScanResult(
-                                vulnerability_type=VulnerabilityType.CODE_PATTERN,
-                                risk_level=func_data["risk_level"],
-                                title=f"Dangerous function: {func_name}()",
-                                description=func_data["description"],
-                                location=f"{file_path}:{node.lineno}",
-                                recommendation=func_data["recommendation"],
-                                details={"function": func_name, "line": node.lineno},
-                            )
-                        )
-
-        except SyntaxError:
-            # 構文エラーファイルはスキップ
-            pass
-        except Exception as e:
-            self.logger.error(f"AST scan error: {e}")
-
-        return results
-
-    def _scan_patterns(self, file_path: Path, content: str) -> List[ScanResult]:
-        """正規表現パターンスキャン"""
-        results: List[ScanResult] = []
-        lines = content.split("\n")
-
-        for pattern_data in self.SECURITY_PATTERNS:
-            pattern = pattern_data["pattern"]
-
-            for line_no, line in enumerate(lines, 1):
-                if re.search(pattern, line):
-                    results.append(
-                        ScanResult(
-                            vulnerability_type=VulnerabilityType.CODE_PATTERN,
-                            risk_level=pattern_data["risk_level"],
-                            title=pattern_data["title"],
-                            description=pattern_data["description"],
-                            location=f"{file_path}:{line_no}",
-                            recommendation=pattern_data["recommendation"],
-                            details={"pattern": pattern, "line_content": line.strip()},
-                        )
-                    )
-
-        return results
-
-    def _get_function_name(self, node: ast.AST) -> str:
-        """関数名取得"""
-        if isinstance(node, ast.Name):
-            return node.id
-        elif isinstance(node, ast.Attribute):
-            return node.attr
-        return ""
-
-
-class RuntimeMonitor:
-    """実行時セキュリティ監視"""
-
-    def __init__(self, config: ScannerConfig, logger: Any):
-        self.config = config
-        self.logger = logger
-        self._monitoring = False
-        self._monitor_thread: Optional[threading.Thread] = None
-        self._suspicious_activities: List[Dict[str, Any]] = []
-
-    def start_monitoring(self) -> None:
-        """監視開始"""
-        if self._monitoring:
-            return
-
-        self._monitoring = True
-        self._monitor_thread = threading.Thread(target=self._monitor_loop, daemon=True)
-        self._monitor_thread.start()
-
-    def stop_monitoring(self) -> None:
-        """監視停止"""
-        self._monitoring = False
-        if self._monitor_thread and self._monitor_thread.is_alive():
-            self._monitor_thread.join(timeout=5)
-
-    def get_security_events(self) -> List[ScanResult]:
-        """セキュリティイベント取得"""
-        results = []
-
-        for activity in self._suspicious_activities:
-            results.append(
-                ScanResult(
-                    vulnerability_type=VulnerabilityType.RUNTIME_BEHAVIOR,
-                    risk_level=activity.get("risk_level", RiskLevel.MEDIUM),
-                    title=activity.get("title", "Suspicious runtime activity"),
-                    description=activity.get("description", ""),
-                    location=activity.get("location", "runtime"),
-                    recommendation=activity.get(
-                        "recommendation", "Investigate the activity"
-                    ),
-                    details=activity,
-                )
-            )
-
-        return results
-
-    def _monitor_loop(self) -> None:
-        """監視メインループ"""
-        while self._monitoring:
-            try:
-                # CPU・メモリ使用量監視
-                self._check_resource_usage()
-
-                # ネットワーク活動監視
-                self._check_network_activity()
-
-                # ファイルシステムアクセス監視
-                self._check_file_access()
-
-                time.sleep(5)  # 5秒間隔
-
-            except Exception as e:
-                self.logger.error(f"Runtime monitoring error: {e}")
-
-    def _check_resource_usage(self) -> None:
-        """リソース使用量チェック"""
-        try:
-            import psutil
-
-            # CPU使用率が90%超過
-            cpu_percent = psutil.cpu_percent(interval=1)
-            if cpu_percent > 90:
-                self._suspicious_activities.append(
-                    {
-                        "title": "High CPU usage detected",
-                        "description": f"CPU usage: {cpu_percent}%",
-                        "risk_level": RiskLevel.MEDIUM,
-                        "timestamp": datetime.now(timezone.utc).isoformat(),
-                        "cpu_percent": cpu_percent,
-                    }
-                )
-
-            # メモリ使用率が85%超過
-            memory = psutil.virtual_memory()
-            if memory.percent > 85:
-                self._suspicious_activities.append(
-                    {
-                        "title": "High memory usage detected",
-                        "description": f"Memory usage: {memory.percent}%",
-                        "risk_level": RiskLevel.MEDIUM,
-                        "timestamp": datetime.now(timezone.utc).isoformat(),
-                        "memory_percent": memory.percent,
-                    }
-                )
-
-        except ImportError:
-            pass
-        except Exception as e:
-            self.logger.error(f"Resource usage check error: {e}")
-
-    def _check_network_activity(self) -> None:
-        """ネットワーク活動チェック"""
-        # 実装簡略化のため基本的なチェックのみ
-        pass
-
-    def _check_file_access(self) -> None:
-        """ファイルアクセスチェック"""
-        # 実装簡略化のため基本的なチェックのみ
-        pass
-
-
-class CVEDatabase:
-    """CVEデータベース管理（簡略版）"""
-
-    def __init__(self, config: ScannerConfig, logger: Any):
-        self.config = config
-        self.logger = logger
-        self._cve_data: Dict[str, List[CVERecord]] = {}
-        self._load_builtin_cve_data()
-
-    def _load_builtin_cve_data(self) -> None:
-        """組み込みCVEデータロード"""
-        # 組み込みCVEデータ（実際の実装では外部データベース使用）
-        self._cve_data = {
-            "django": [
-                CVERecord(
-                    cve_id="CVE-2022-34265",
-                    description="Django SQL injection vulnerability",
-                    severity="HIGH",
-                    score=8.8,
-                    affected_versions=["<3.2.14", "<4.0.6"],
-                    fixed_in="3.2.14, 4.0.6",
-                    published_date="2022-07-04",
-                )
-            ],
-            "pillow": [
-                CVERecord(
-                    cve_id="CVE-2022-22817",
-                    description="Pillow buffer overflow vulnerability",
-                    severity="HIGH",
-                    score=9.8,
-                    affected_versions=["<9.0.1"],
-                    fixed_in="9.0.1",
-                    published_date="2022-01-10",
-                )
-            ],
-        }
-
-    def lookup_cve(self, package_name: str, version: str) -> List[CVERecord]:
-        """CVE検索"""
-        results = []
-        package_name = package_name.lower()
-
-        if package_name in self._cve_data:
-            for cve in self._cve_data[package_name]:
-                # バージョンマッチング（簡略化）
-                if self._version_affected(version, cve.affected_versions):
-                    results.append(cve)
-
-        return results
-
-    def _version_affected(self, version: str, affected_versions: List[str]) -> bool:
-        """バージョン影響チェック"""
-        for affected in affected_versions:
-            try:
-                if affected.startswith("<"):
-                    if version < affected[1:]:
-                        return True
-                elif affected.startswith("<="):
-                    if version <= affected[2:]:
-                        return True
-                elif version == affected:
-                    return True
-            except Exception:
-                continue
-        return False
+# 分離されたコンポーネントのインポート（後方互換性完全確保）
+from .code_pattern_scanner import CodePatternScanner
+from .cve_database import CVEDatabase
+from .dependency_scanner import DependencyScanner
+from .runtime_monitor import RuntimeMonitor
+from .vuln_types import (
+    CVERecord,
+    RiskLevel,
+    ScanResult,
+    ScannerConfig,
+    SecurityReport,
+    VulnerabilityType,
+)
 
 
 class VulnerabilityScanner:
@@ -936,6 +332,28 @@ def quick_scan(file_or_dir: Union[str, Path]) -> List[ScanResult]:
         return scanner.code_pattern_scanner.scan_file(path)
     else:
         return scanner.scan_code_patterns(path)
+
+
+# 既存APIの完全再現（後方互換性100%保持）
+__all__ = [
+    # 基本型・データクラス（vuln_types.pyから再エクスポート）
+    "RiskLevel",
+    "VulnerabilityType",
+    "CVERecord",
+    "ScanResult",
+    "SecurityReport",
+    "ScannerConfig",
+    # 分離されたコンポーネント
+    "DependencyScanner",
+    "CodePatternScanner",
+    "RuntimeMonitor",
+    "CVEDatabase",
+    # メインクラス
+    "VulnerabilityScanner",
+    # コンビニエンス関数
+    "scan_project",
+    "quick_scan",
+]
 
 
 if __name__ == "__main__":
