+++ kumihan_formatter/parsers/block/handlers.py
@@ -27,7 +27,8 @@ from .utils import (
 )
 
 if TYPE_CHECKING:
-    from ..keyword.keyword_parser import KeywordParser
+    # 統合済みパーサーを使用 - Issue #1168 Parser Responsibility Separation
+    from ...core.parsing.specialized.keyword_parser import UnifiedKeywordParser as KeywordParser
 
 
 class BlockHandler:
