+++ kumihan_formatter/parsers/block/parser.py
@@ -29,7 +29,8 @@ from .utils import (
 )
 
 if TYPE_CHECKING:
-    from ..keyword.keyword_parser import KeywordParser
+    # 統合済みパーサーを使用 - Issue #1168 Parser Responsibility Separation
+    from ...core.parsing.specialized.keyword_parser import UnifiedKeywordParser as KeywordParser
 
 
 class UnifiedBlockParser:
