--- kumihan_formatter/parsers/block_parser_legacy.py
@@ -1,4 +1,0 @@
-# Legacy wrapper
-from .block.block_parser import UnifiedBlockParser
-
-BlockParser = UnifiedBlockParser
