--- kumihan_formatter/parsers/keyword_parser_legacy.py
@@ -1,4 +1,0 @@
-# Legacy wrapper
-from .keyword.keyword_parser import UnifiedKeywordParser
-
-KeywordParser = UnifiedKeywordParser
